# Api-Backend-LocSystem/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LocSystem API",
        "description": "Documentação da API do sistema LocSystem",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin-users": {
            "get": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Lista os administradores cadastrados com filtros por busca e periodo.",
                "operationId": "577ee0bbbdcbbe057fb59b9ce782eb64",
                "responses": {
                    "200": {
                        "description": "Administradores retornados com sucesso no campo adminUsers."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem consultar administradores."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao listar os administradores."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Cadastra um novo administrador com acesso inicial ao sistema.",
                "operationId": "2797df51dc649387ebccfcb22fd0d31e",
                "responses": {
                    "201": {
                        "description": "Administrador cadastrado com sucesso e retornado no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado. Apenas administradores podem cadastrar novos administradores."
                    },
                    "409": {
                        "description": "O e-mail informado ja esta em uso por outro usuario."
                    },
                    "422": {
                        "description": "Erro de validacao nos dados enviados para cadastro do administrador."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao cadastrar o administrador."
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 20
                                    },
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 32
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_phone",
                                    "v_password"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-users/{id}": {
            "get": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Retorna os detalhes de um administrador especifico.",
                "operationId": "7348205e3a6f34152b5b04746d49477b",
                "responses": {
                    "200": {
                        "description": "Administrador retornado com sucesso no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao consultar o administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Atualiza os dados cadastrais de um administrador.",
                "operationId": "364a4cd8e3c19bfea38ea286efb10bc2",
                "responses": {
                    "200": {
                        "description": "Administrador atualizado com sucesso e retornado no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "409": {
                        "description": "Conflito de e-mail ou tentativa de gerenciar o administrador principal."
                    },
                    "422": {
                        "description": "Erro de validacao nos dados enviados para atualizacao do administrador."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_name": {
                                        "type": "string",
                                        "minLength": 3,
                                        "maxLength": 255
                                    },
                                    "v_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "v_phone": {
                                        "type": "string",
                                        "minLength": 10,
                                        "maxLength": 20
                                    }
                                },
                                "required": [
                                    "v_name",
                                    "v_email",
                                    "v_phone"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Remove um administrador e seus vinculos de autenticacao relacionados.",
                "operationId": "5cc62e6a7bcaef915bf7146939a5f730",
                "responses": {
                    "200": {
                        "description": "Administrador removido com sucesso e confirmacao retornada no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "409": {
                        "description": "Nao e permitido excluir o proprio administrador autenticado nem gerenciar o administrador principal nesta acao."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao remover o administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-users/{id}/approval": {
            "patch": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Aprova um administrador pendente para acesso ao sistema.",
                "operationId": "8e1e03216b248437460e12b19969c3ce",
                "responses": {
                    "200": {
                        "description": "Administrador aprovado com sucesso e retornado no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "409": {
                        "description": "Nao e permitido aprovar o administrador principal nesta acao."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao aprovar o administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-users/{id}/status": {
            "patch": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Bloqueia ou desbloqueia um administrador conforme o status informado.",
                "operationId": "a51bddf0c51f1bbc582a8369bfc4dce4",
                "responses": {
                    "200": {
                        "description": "Status do administrador atualizado com sucesso e dados retornados no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "409": {
                        "description": "Nao e permitido bloquear o proprio administrador autenticado nem gerenciar o administrador principal nesta acao."
                    },
                    "422": {
                        "description": "Erro de validacao no payload de alteracao de status do administrador."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao atualizar o status do administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isActive": {
                                        "type": "boolean"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 1000
                                    }
                                },
                                "required": [
                                    "isActive"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-users/{id}/password": {
            "patch": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Define uma nova senha para o administrador informado.",
                "operationId": "8cc728fc8c3dd7550d624c3d33c25d7b",
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso e confirmacao retornada no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado, administrador nao encontrado ou conta de autenticacao inexistente."
                    },
                    "422": {
                        "description": "Erro de validacao da nova senha informada."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao alterar a senha do administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "v_password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "maxLength": 32
                                    }
                                },
                                "required": [
                                    "v_password"
                                ]
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-users/{id}/reset-2fa": {
            "post": {
                "tags": [
                    "Administradores"
                ],
                "summary": "Reseta a configuracao de autenticacao em dois fatores de um administrador.",
                "operationId": "51a65762ae9bdc42c3a116baa1dd69be",
                "responses": {
                    "200": {
                        "description": "2FA resetado com sucesso e confirmacao retornada no campo adminUser."
                    },
                    "401": {
                        "description": "Acesso negado ou administrador nao encontrado."
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao resetar o 2FA do administrador."
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Parametro de rota id.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}